+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Remove chat-plugins
   * Remove static folder
   * Move lib/ to config/
+  * Move cidr.js and dnsbl.js to config folder
 
 1.1.0 / 2015-03-15
 ==================